+++ package.json
@@ -10,9 +10,13 @@
     "@nestjs/axios": "^3.0.1",
     "@nestjs/cli": "^10.3.0",
     "axios": "^1.6.6",
+    "class-transformer": "^0.5.1",
     "concurrently": "^8.2.2",
     "framer-motion": "^10.18.0",
     "react-router-dom": "^6.21.3",
     "react-toastify": "^10.0.4"
+  },
+  "devDependencies": {
+    "@types/sanitize-html": "^2.11.0"
   }
 }
