+++ server/src/auth/auth.controller.ts
@@ -5,10 +5,13 @@ import {
   Get,
   UseGuards,
   Request,
+  Req,
 } from '@nestjs/common';
 import { CreateUserDto } from 'src/user/dto/create-user.dto';
 import { AuthService } from './auth.service';
 import { AuthGuard } from './auth.guard';
+import { CreateTodoDto } from './dto/create-todo.dto';
+// import { TodoService } from 'src/todo/todo.service';
 
 type LoginDTO = {
   username: string;
@@ -51,5 +54,13 @@ export class AuthController {
     } else {
       return 'No User';
     }
+  }
+
+  @UseGuards(AuthGuard)
+  @Post('/create-todo')
+  async createTodo(@Body() createTodoDto: CreateTodoDto, @Request() req) {
+    console.log(createTodoDto);
+    console.log('USER REQUEST', req.sub);
+    // return this.authService.createTodo(createTodoDto);
   }
 }
