+++ server/src/todo/todo.module.ts
@@ -1,6 +1,5 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { TodoService } from './todo.service';
-import { TodoController } from './todo.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Todo } from './entities/todo.entity';
 import { UserModule } from 'src/user/user.module';
