+++ server/src/todo/todo.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { CreateTodoDto } from './dto/create-todo.dto';
-import { UpdateTodoDto } from './dto/update-todo.dto';
+import { CreateTodoDto } from '../auth/dto/create-todo.dto';
+import { UpdateTodoDto } from '../auth/dto/update-todo.dto';
 import { Todo } from './entities/todo.entity';
 
 @Injectable()
@@ -11,12 +11,18 @@ export class TodoService {
     @InjectRepository(Todo) private todoRepository: Repository<Todo>,
   ) {}
 
-  createTodo(createTodoDto: CreateTodoDto): Promise<Todo> {
-    const newTodo: Todo = new Todo();
-    newTodo.todo = createTodoDto.todo;
-    newTodo.reflectionText = createTodoDto.reflectionText;
-    newTodo.priority = createTodoDto.priority;
-    return this.todoRepository.save(newTodo);
+  async findUserProjects(id: number): Promise<Todo[]> {
+    return await this.todoRepository.find({ where: { user: { id } } });
+  }
+
+  async createTodo(createTodoDto: CreateTodoDto) {
+    // await this.todoRepository.save({
+    //   todo: createTodoDto.todo,
+    //   reflectionText: createTodoDto.reflectionText,
+    //   priority: createTodoDto.priority,
+    //   userId: createTodoDto.id,
+    // });
+    // return this.findUserProjects(createTodoDto.id);
   }
 
   findAllTodos(): Promise<Todo[]> {
